DP-1169: Moved datasetDetails, datasetList to module
