+++ src/modules/DatasetList/Home.vue
@@ -65,7 +65,7 @@
 import { mapGetters, mapActions, mapState } from 'vuex'
 import IconDotDotHorizontal from '@/components/icons/IconDotDotHorizontal'
 import ExpandableRow from '@/components/ExpandableRow'
-import DatasetStatus from '@/components/DatasetList/DatasetStatus'
+import DatasetStatus from '@/modules/DatasetList/DatasetStatus'
 import datasets from '@/assets/mock/dataset.json'
 
 export default {
